// packages/providers/src/pbyp/directus-schema.ts
export const PBYP_SCHEMA: PbypSchema = {
  address: {
    primaryKey: "id",
    fields: {
      access_entities: {
        type: "alias",
        readonly: true,
        relation: {
          kind: "m2m",
          through: "address_entities",
          fk: "address_id",
          otherFk: "entities_id",
          to: "entities",
        },
      },
      address: {
        type: "string",
      },
      address_book: {
        type: "integer",
        relation: {
          kind: "m2o",
          to: "address_book",
        },
      },
      city: {
        type: "string",
      },
      code: {
        type: "string",
        required: true,
      },
// ...
    },
  },
// ...
  entities: {
    primaryKey: "id",
    fields: {
      address_book: {
        type: "integer",
        relation: {
          kind: "m2o",
          to: "address_book",
        },
      },
      associated_entities: {
        type: "alias",
        relation: {
          kind: "m2m",
          through: "associated_entities",
          fk: "entities_id",
          otherFk: "related_entities_id",
          to: "entities",
        },
      },
      association_key: {
        type: "string",
      },
// ...
    },
  },
// ...
};